requirements.txt
# Essential dependencies for basic functionality
requests>=2.31.0
pandas>=2.1.0
numpy>=1.24.0
pydantic>=2.5.0
rich>=13.7.0
typer>=0.9.0
loguru>=0.7.0
pyyaml>=6.0.1

# File operations
openpyxl>=3.1.0

# Web scraping
beautifulsoup4>=4.12.0

# System monitoring
psutil>=5.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional AI dependencies (install separately if needed)
# torch>=2.0.0
# transformers>=4.35.0
# accelerate>=0.24.0
